scrapers/requirements.txt
anyio==3.6.2
attrs==22.2.0
Automat==22.10.0
beautifulsoup4==4.11.2
billiard==3.6.4.0
blobfile==2.0.1
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==39.0.2
cssselect==1.2.0
filelock==3.9.0
firebase-admin==6.1.0
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-pubsub==2.15.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.3
grpcio-status==1.51.3
h11==0.14.0
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
lxml==4.9.2
msgpack==1.0.5
OpenAIAuth==0.3.2
packaging==23.0
parsel==1.7.0
prompt-toolkit==3.0.38
Protego==0.2.1
proto-plus==1.22.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.17
PyDispatcher==2.0.7
PyJWT==2.6.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
PySocks==1.7.1
queuelib==1.6.2
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
revChatGPT==3.1.5
rfc3986==1.5.0
rsa==4.9
Scrapy==2.8.0
scrapyscript==1.1.5
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.4
tiktoken==0.3.0
tldextract==3.4.0
Twisted==22.10.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
w3lib==2.1.1
wcwidth==0.2.6
zope.interface==5.5.2
